+++ Sources/Controls/DualArcKnob.swift
@@ -1,11 +1,11 @@
 import SwiftUI
 
 public struct DualArcKnob: View {
-    @Binding var volume: Double
-    @Binding var pan: Double
+    @Binding var volume: Float
+    @Binding var pan: Float
     var rangeDegrees = 270.0
 
-    public init(volume: Binding<Double>, pan: Binding<Double>) {
+    public init(volume: Binding<Float>, pan: Binding<Float>) {
         _volume = volume
         _pan = pan
     }
@@ -24,7 +24,7 @@ public struct DualArcKnob: View {
 
                 // Stroke value trim of knob
                 Circle()
-                    .trim(from: 45 / 360.0, to: (45 + volume * rangeDegrees) / 360.0)
+                    .trim(from: 45 / 360.0, to: (45 + Double(volume) * rangeDegrees) / 360.0)
                     .rotation(.degrees(-rangeDegrees))
                     .stroke(.red, style: StrokeStyle(lineWidth: geo.size.width / 10, lineCap: .round))
                     .frame(width: geo.size.width * 0.8, height: geo.size.height * 0.8)
@@ -38,7 +38,7 @@ public struct DualArcKnob: View {
 
                 // Stroke value trim of knob
                 Circle()
-                    .trim(from: 45 / 360.0, to: (45 + pan * rangeDegrees) / 360.0)
+                    .trim(from: 45 / 360.0, to: (45 + Double(pan) * rangeDegrees) / 360.0)
                     .rotation(.degrees(-rangeDegrees))
                     .stroke(.red, style: StrokeStyle(lineWidth: geo.size.width / 10, lineCap: .round))
                     .frame(width: geo.size.width * 0.5, height: geo.size.height * 0.5)
